Keep logic in `useDepositBTCTransaction`
