+++ dapp/src/components/TransactionModal/ActiveStakingStep/DepositBTCModal.tsx
@@ -3,11 +3,10 @@ import {
   useActionFlowPause,
   useActionFlowTokenAmount,
   useAppDispatch,
-  useConnector,
+  useDepositBTCTransaction,
   useInvalidateQueries,
   useStakeFlowContext,
   useVerifyDepositAddress,
-  useWallet,
 } from "#/hooks"
 import { eip1193, logPromiseFailure } from "#/utils"
 import { PROCESS_STATUSES } from "#/types"
@@ -20,8 +19,6 @@ import WalletInteractionModal from "../WalletInteractionModal"
 const { userKeys } = queryKeysFactory
 
 export default function DepositBTCModal() {
-  const connector = useConnector()
-  const { address } = useWallet()
   const tokenAmount = useActionFlowTokenAmount()
   const { btcAddress, depositReceipt, stake } = useStakeFlowContext()
   const verifyDepositAddress = useVerifyDepositAddress()
@@ -70,32 +67,7 @@ export default function DepositBTCModal() {
     [onError, handlePause],
   )
 
-  const { mutate: sendBitcoinTransaction, status } = useMutation({
-    mutationKey: ["send-bitcoin"],
-    mutationFn: async ({
-      recipient,
-      amount,
-    }: {
-      recipient: string
-      amount: bigint
-    }) => {
-      if (!address) {
-        throw new Error("Bitcoin account was not connected.")
-      }
-
-      if (!connector) {
-        throw new Error("Connector was not defined.")
-      }
-      // @ts-expect-error adjust types to handle bitcoin wallet wrappers
-      const client = await connector.getClient()
-
-      const satoshis = Number(amount)
-
-      // @ts-expect-error adjust types to handle bitcoin wallet wrappers
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-assignment
-      const txhash: string = await client.sendBitcoin(recipient, satoshis)
-      return txhash
-    },
+  const { mutate: sendBitcoinTransaction, status } = useDepositBTCTransaction({
     onSuccess: onDepositBTCSuccess,
     onError: onDepositBTCError,
   })
